+++ sn3218/__init__.py
@@ -1,6 +1,4 @@
-from smbus import SMBus
-
-__version__ = '2.0.0'
+__version__ = '3.0.0'
 
 
 I2C_ADDRESS = 0x54
@@ -14,15 +12,9 @@ _sn3218 = None  # For module compatibiity shim
 
 
 class SN3218:
-    def __init__(self, i2c_bus=None, i2c_dev=None, enable_mask=0b111111111111111111):
+    def __init__(self, i2c_bus=1, i2c_dev=None, enable_mask=0b111111111111111111):
         if i2c_dev is None:
-            if i2c_bus is None:
-                import RPi.GPIO as GPIO
-                if GPIO.RPI_REVISION < 2:
-                    i2c_bus = 0
-                else:
-                    i2c_bus = 1
-
+            from smbus2 import SMBus
             self.i2c = SMBus(i2c_bus)
         else:
             self.i2c = i2c_dev
@@ -142,7 +134,7 @@ def output_raw(values):
 
 
 def channel_gamma(channel, gamma_table):
-    _get_sn3218().enable()
+    _get_sn3218().channel_gamma(channel, gamma_table)
 
 
 def enable_leds(enable_mask):