+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+3.0.0
+-----
+
+* BREAKING: Remove RPi.GPIO and bus autodetect
+* BREAKING: Switch to smbus2
+
 2.0.0
 -----
 
@@ -38,7 +44,7 @@
 1.2.1
 -----
 
-* Removed hard dependancy upon SMbus, added message at runtime
+* Removed hard dependency upon SMbus, added message at runtime
 
 1.0.0
 -----